requirements.txt
playwright==1.40.0
click==8.1.7
aiofiles==23.2.1
pydantic==2.5.1
python-dotenv==1.0.0
openai-1.106.1
rich==13.7.0
beautifulsoup4==4.12.2